separate beans package so clients can use it
