+++ galasa-parent/dev.galasa.framework.api.beans/bnd.bnd
@@ -1,0 +1,3 @@
+-snapshot ${tstamp}
+Bundle-Name: Galasa REST API beans
+Export-Package: dev.galasa.framework.api.beans
