+++ galasa-parent/dev.galasa.framework.api.beans/build.gradle
@@ -10,15 +10,6 @@ description = 'Framework API - Common Packages'
 version = '0.33.0'
 
 dependencies {
-    implementation project(':dev.galasa.framework')
-    implementation project(':dev.galasa.framework.api.beans')
-
-    testFixturesImplementation 'javax.servlet:javax.servlet-api:3.1.0'
-    testFixturesImplementation 'org.assertj:assertj-core:3.16.1'
-    testFixturesImplementation 'com.google.code.gson:gson:2.10.1'
-    testFixturesImplementation(project(':dev.galasa.framework'))
-    testFixturesImplementation 'javax.validation:validation-api:2.0.1.Final'
-
 }
 
 // Note: These values are consumed by the parent build process
@@ -28,9 +19,9 @@ dependencies {
 // which gathers-up all the packaging metadata about all the OSGi bundles in this component.
 ext.projectName=project.name
 ext.includeInOBR          = true
-ext.includeInMVP          = false
+ext.includeInMVP          = true
 ext.includeInIsolated     = true
-ext.includeInBOM          = false
+ext.includeInBOM          = true
 ext.includeInCodeCoverage = true
 ext.includeInJavadoc      = true
