+++ galasa-parent/dev.galasa.framework.api.beans/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = "dev.galasa.framework.api.beans"
