+++ galasa-parent/dev.galasa.framework.api.beans/src/main/java/dev/galasa/framework/api/beans/GalasaProperty.java
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package dev.galasa.framework.api.common.resources.beans;
+package dev.galasa.framework.api.beans;
 
 public class GalasaProperty {
