--- galasa-parent/dev.galasa.framework.api.client/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = "dev.galasa.framework.api.client"
