+++ galasa-parent/dev.galasa.framework.api.common/bnd.bnd
@@ -1,4 +1,4 @@
 -snapshot ${tstamp}
 Bundle-Name: Galasa API - Common Packages
 Export-Package: dev.galasa.framework.api.common*
-Import-Package: *
+Import-Package: dev.galasa.framework.api.beans,*
