+++ galasa-parent/dev.galasa.framework.api.common/src/test/java/dev/galasa/framework/api/common/resources/TestGalasaProperty.java
@@ -7,8 +7,8 @@ package dev.galasa.framework.api.common.resources;
 
 import org.junit.Test;
 
+import dev.galasa.framework.api.beans.GalasaProperty;
 import dev.galasa.framework.api.common.InternalServletException;
-import dev.galasa.framework.api.common.resources.beans.GalasaProperty;
 
 import static org.assertj.core.api.Assertions.*;
