+++ galasa-parent/dev.galasa.framework.api.cps/bnd.bnd
@@ -2,6 +2,7 @@
 Bundle-Name: Galasa API Configuration Property Store microservices
 Export-Package: dev.galasa.framework.api.cps*;
 Import-Package: dev.galasa.framework.api.common*, \
+                dev.galasa.framework.api.beans, \
                 dev.galasa.framework.api.authentication, \
                 *
