+++ galasa-parent/dev.galasa.framework.api.cps/build.gradle
@@ -10,6 +10,7 @@ version = '0.32.0'
 
 dependencies {
     implementation project(':dev.galasa.framework')
+    implementation project(':dev.galasa.framework.api.beans')
     implementation project(':dev.galasa.framework.api.common')
     testImplementation(testFixtures(project(':dev.galasa.framework.api.common')))
     testImplementation 'org.assertj:assertj-core:3.23.1'
