+++ galasa-parent/dev.galasa.framework.api.cps/src/main/java/dev/galasa/framework/api/cps/internal/routes/CPSRoute.java
@@ -15,6 +15,9 @@ import java.util.Map;
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.constraints.NotNull;
 
+import dev.galasa.framework.api.beans.GalasaProperty;
+import dev.galasa.framework.api.beans.GalasaPropertyData;
+import dev.galasa.framework.api.beans.GalasaPropertyMetadata;
 import dev.galasa.framework.api.common.BaseRoute;
 import dev.galasa.framework.api.common.InternalServletException;
 import dev.galasa.framework.api.common.ResponseBuilder;
@@ -25,9 +28,6 @@ import dev.galasa.framework.api.common.resources.CPSProperty;
 import dev.galasa.framework.api.common.resources.GalasaPropertyName;
 import dev.galasa.framework.api.common.resources.ResourceNameValidator;
 import dev.galasa.framework.api.common.resources.beans.GalasaBeanSerialiser;
-import dev.galasa.framework.api.common.resources.beans.GalasaProperty;
-import dev.galasa.framework.api.common.resources.beans.GalasaPropertyMetadata;
-import dev.galasa.framework.api.common.resources.beans.GalasaPropertyData;
 import dev.galasa.framework.spi.ConfigurationPropertyStoreException;
 import dev.galasa.framework.spi.FrameworkException;
 import dev.galasa.framework.spi.IFramework;
