+++ galasa-parent/dev.galasa.framework.api.resources/build.gradle
@@ -13,6 +13,7 @@ version '0.32.0'
 
 dependencies {
     implementation project(':dev.galasa.framework')
+    implementation project(':dev.galasa.framework.api.beans')
     implementation project(':dev.galasa.framework.api.common')
     implementation project(':dev.galasa.framework.api.cps')
