+++ galasa-parent/settings.gradle
@@ -22,7 +22,7 @@ include 'dev.galasa.framework.metrics'
 include 'dev.galasa.framework.api'
 include 'dev.galasa.framework.api.authentication'
 include 'dev.galasa.framework.api.bootstrap'
-include 'dev.galasa.framework.api.client'
+include 'dev.galasa.framework.api.beans'
 include 'dev.galasa.framework.api.common'
 include 'dev.galasa.framework.api.cps'
 include 'dev.galasa.framework.api.health'
