+++ release.yaml
@@ -146,8 +146,8 @@ api:
     obr: true
     isolated: true
 
-  - artifact: dev.galasa.framework.api.client
-    version: 0.32.0
+  - artifact: dev.galasa.framework.api.beans
+    version: 0.33.0
     obr: true
     bom: true
     mvp: true
